requirements.txt
openai
requests
beautifulsoup4
pyyaml
ddgs
pydantic
pytest
ruff
mypy